Make nginx listen on both the IPv4 and IPv6 wildcard addresses
The nginx role always assumes that Open edX services listen on
the wildcard address (we can only ever configure ports, not listen
addresses). However, if nginx is configured with a "listen" directive
that contains just a port number, it will listen only on the IPv4
wildcard address, 0.0.0.0.

For IPv6-enabled hosts, that means that Open edX service will not
be accessible through the system's IPv6 address or AAAA hostname.

Amend the listen directive to instead use the IPv6 wildcard address,
"[::]", and add the ipv6only=off option, which makes nginx also
listen on the IPv4 wildcard when so configured.